+++ app/src/androidTest/java/ch/epfl/sweng/zuluzulu/fragments/ProfileFragmentTest.java
@@ -45,7 +45,6 @@ public class ProfileFragmentTest {
     @Rule
     public GrantPermissionRule permissionRule = GrantPermissionRule.grant(Manifest.permission.CAMERA);
 
-
     @Rule
     public GrantPermissionRule permissionRule1 = GrantPermissionRule.grant(Manifest.permission.READ_EXTERNAL_STORAGE);
