+++ app/src/main/res/layout/fragment_main.xml
@@ -56,7 +56,7 @@
                             android:layout_height="wrap_content" />
                     </LinearLayout>
 
-                    <ch.epfl.sweng.zuluzulu.utility.MyListView
+                    <ch.epfl.sweng.zuluzulu.utility.NonScrollableListView
                         android:id="@+id/main_page_list_event"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
@@ -102,7 +102,7 @@
                             android:layout_height="wrap_content" />
                     </LinearLayout>
 
-                    <ch.epfl.sweng.zuluzulu.utility.MyListView
+                    <ch.epfl.sweng.zuluzulu.utility.NonScrollableListView
                         android:id="@+id/main_page_random_assos"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
@@ -112,7 +112,7 @@
                         tools:layout_editor_absoluteX="8dp">
 
 
-                    </ch.epfl.sweng.zuluzulu.utility.MyListView>
+                    </ch.epfl.sweng.zuluzulu.utility.NonScrollableListView>
                 </LinearLayout>
 
                 <Button
