+++ app/src/test/java/ch/epfl/sweng/zuluzulu/utility/UtilsTest.java
@@ -10,6 +10,9 @@ import org.junit.runners.JUnit4;
 
 import java.util.Date;
 
+import static ch.epfl.sweng.zuluzulu.utility.GeopointUtility.distanceBetween;
+import static ch.epfl.sweng.zuluzulu.utility.GeopointUtility.toGeoPoint;
+import static ch.epfl.sweng.zuluzulu.utility.TimeUtility.getMillisecondsSince;
 import static ch.epfl.sweng.zuluzulu.utility.Utils.*;
 import static junit.framework.TestCase.assertEquals;
 import static junit.framework.TestCase.assertFalse;
@@ -80,32 +83,32 @@ public class UtilsTest {
                     strMinute = "0" + strMinute;
                 }
                 String expected = strHour + ":" + strMinute;
-                assertEquals(expected, Utils.hourAndMinutesFrom(hour, minute));
+                assertEquals(expected, TimeUtility.hourAndMinutesFrom(hour, minute));
             }
         }
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testIllegalHourInHourAndMinutes() {
-        Utils.hourAndMinutesFrom(-2, 45);
+        TimeUtility.hourAndMinutesFrom(-2, 45);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testIllegalMinuteInHourAndMinutes() {
-        Utils.hourAndMinutesFrom(12, 103);
+        TimeUtility.hourAndMinutesFrom(12, 103);
     }
 
     @Test
     public void testValidHour() {
-        assertTrue(Utils.validHour(12));
-        assertFalse(Utils.validHour(-2));
-        assertFalse(Utils.validHour(24));
+        assertTrue(TimeUtility.validHour(12));
+        assertFalse(TimeUtility.validHour(-2));
+        assertFalse(TimeUtility.validHour(24));
     }
 
     @Test
     public void testValidMinute() {
-        assertTrue(Utils.validMinute(50));
-        assertFalse(Utils.validMinute(-1));
-        assertFalse(Utils.validMinute(60));
+        assertTrue(TimeUtility.validMinute(50));
+        assertFalse(TimeUtility.validMinute(-1));
+        assertFalse(TimeUtility.validMinute(60));
     }
 }
